+++ CMakeLists.txt
@@ -8,37 +8,12 @@ set(CMAKE_CXX_STANDARD 11)
 # TODO check flags to use
 set(CMAKE_CXX_FLAGS "-Wall -ggdb -O3 -fPIC -march=native -mno-avx")
 
-# Download and unpack googletest at configure time
-configure_file(CMakeLists.txt.in googletest-download/CMakeLists.txt)
-execute_process(COMMAND ${CMAKE_COMMAND} -G "${CMAKE_GENERATOR}" .
-  RESULT_VARIABLE result
-  WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/googletest-download )
-if(result)
-  message(FATAL_ERROR "CMake step for googletest failed: ${result}")
+include_directories("include")
+
+add_subdirectory(src)
+
+option(RUNIN_ENABLE_TESTS "Enable tests" OFF)
+
+if (RUNIN_ENABLE_TESTS)
+    add_subdirectory(test)
 endif()
-execute_process(COMMAND ${CMAKE_COMMAND} --build .
-  RESULT_VARIABLE result
-  WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/googletest-download )
-if(result)
-  message(FATAL_ERROR "Build step for googletest failed: ${result}")
-endif()
-
-# Add googletest directly to our build. This defines
-# the gtest and gtest_main targets.
-add_subdirectory(${CMAKE_CURRENT_BINARY_DIR}/googletest-src
-                 ${CMAKE_CURRENT_BINARY_DIR}/googletest-build
-                 EXCLUDE_FROM_ALL)
-
-include_directories("${gtest_SOURCE_DIR}/include")
-
-
- 
-# Runin
-file(GLOB SRC_FILES ${PROJECT_SOURCE_DIR}/*.cpp)
-
-add_library(runTests_lib ${SRC_FILES})
-add_executable(runTests test_phys.cpp)
-
-target_link_libraries(runTests gtest_main pthread)
-target_link_libraries(runTests runTests_lib)
-enable_testing()
