+++ src/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SOURCE_FILES
+    critical_field.cpp
+    growth_rate.cpp
+)
+
+add_library(Runin ${SOURCE_FILES})
